_August 25/Task 2/requirements.txt
pandas
numpy
azure-storage-blob
